S3/S3A1-ContentSummarizer/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
youtube-transcript-api
beautifulsoup4
requests
pypdf
jinja2
